Fiz: corrige o metodo delete de veiculo.

+++ src/main/java/com/rentalCar/managementBackend/controllers/VehicleController.java
@@ -1,17 +1,12 @@
 package com.rentalCar.managementBackend.controllers;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import com.rentalCar.managementBackend.entitys.VehicleEntity;
 import com.rentalCar.managementBackend.services.VehicleService;
@@ -71,10 +66,15 @@ public class VehicleController {
 
     @Operation(summary = "Exclui o veiculo que corresponde ao id informado.")
     @DeleteMapping
-    public ResponseEntity<HttpStatus> deleteVehicle(@PathParam("id") Integer id){
+    public ResponseEntity<String> deleteVehicle(@PathParam("id") Integer id){
         try {
-            vehicleService.deleteVehicleById(id);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            Optional<VehicleEntity> vehicle = vehicleService.getVehicleById(id);
+            if(vehicle.get().getReserved() == false) {
+                vehicleService.deleteVehicleById(id);
+                return new ResponseEntity<>("Veículo excluído com sucesso!!!", HttpStatus.NO_CONTENT);
+            } else {
+                return new ResponseEntity<>("Veículo não pode ser excluido pois está reservado.", HttpStatus.CONFLICT);
+            }
         } catch (Exception e) {
             return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
         }
